# frontend/openapi.json
{
  "components" : {
    "schemas" : {
...
      "ContextRequest" : {
        "properties" : {
          "maxChars" : {
            "format" : "int32",
            "type" : "integer"
          },
          "query" : {
            "type" : "string"
          },
          "scope" : {
            "type" : "string"
          },
          "topK" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "ContextResult" : {
        "properties" : {
          "candidatesConsidered" : {
            "format" : "int32",
            "type" : "integer"
          },
          "context" : {
            "type" : "string"
          },
          "usedMemoryIds" : {
            "items" : {
              "format" : "int64",
              "type" : "integer"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
...
      "GatewayRequestLogEntity" : {
        "properties" : {
          "chosenModel" : {
            "type" : "string"
          },
          "chosenProvider" : {
            "type" : "string"
          },
          "complexity" : {
            "format" : "double",
            "type" : "number"
          },
          "costUsd" : {
            "format" : "double",
            "type" : "number"
          },
          "createdAt" : {
            "description" : "Epoch seconds, possibly fractional",
            "type" : "number"
          },
          "developerId" : {
            "type" : "string"
          },
          "failoverCount" : {
            "format" : "int32",
            "type" : "integer"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "latencyMs" : {
            "format" : "int64",
            "type" : "integer"
          },
          "requestedModel" : {
            "type" : "string"
          },
          "routingReason" : {
            "type" : "string"
          },
          "success" : {
            "type" : "boolean"
          },
          "tokens" : {
            "format" : "int32",
            "type" : "integer"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "InboundResult" : {
        "properties" : {
          "blocked" : {
            "type" : "boolean"
          },
          "injectionHits" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "injectionScore" : {
            "format" : "double",
            "type" : "number"
          },
          "redactions" : {
            "items" : {
              "$ref" : "#/components/schemas/Match"
            },
            "type" : "array"
          },
          "sanitized" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "Match" : {
        "properties" : {
          "category" : {
            "type" : "string"
          },
          "count" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "RunRequest" : {
        "properties" : {
          "input" : {
            "additionalProperties" : {
              "type" : "object"
            },
            "type" : "object"
          },
          "workflowId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "TextRequest" : {
        "properties" : {
          "text" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    }
  },
  "info" : {
    "description" : "Durable execution engine for AI agents: workflows, gateway and console APIs.",
    "title" : "Continuum",
    "version" : "1"
  },
  "openapi" : "3.0.1",
  "paths" : {
...
    "/api/gateway/pipeline/{name}" : {
      "post" : {
        "operationId" : "run_2",
        "parameters" : [ {
          "in" : "path",
          "name" : "name",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RunRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "pipeline-gateway-controller" ]
      }
    },
    "/api/gateway/pipeline/{name}/upload" : {
      "post" : {
        "operationId" : "runUpload_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "name",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "prompt",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "input",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "properties" : {
                  "file" : {
                    "format" : "binary",
                    "type" : "string"
                  }
                },
                "required" : [ "file" ],
                "type" : "object"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "pipeline-gateway-controller" ]
      }
    },
    "/api/gateway/requests" : {
      "get" : {
        "operationId" : "requests_2",
        "parameters" : [ {
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "default" : 100,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/GatewayRequestLogEntity"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "gateway-stats-controller" ]
      }
    },
...
    "/api/memory/context" : {
      "post" : {
        "operationId" : "context",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ContextRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ContextResult"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "memory-controller" ]
      }
    },
...
    "/api/portal/developer/demo/firewall" : {
      "post" : {
        "operationId" : "firewall_1",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TextRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/InboundResult"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "demo-controller" ]
      }
    },
...
    "/api/portal/developer/pipelines/{name}/run/upload" : {
      "post" : {
        "operationId" : "runUpload",
        "parameters" : [ {
          "in" : "path",
          "name" : "name",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "prompt",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "input",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "properties" : {
                  "file" : {
                    "format" : "binary",
                    "type" : "string"
                  }
                },
                "required" : [ "file" ],
                "type" : "object"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : {
                    "type" : "object"
                  },
                  "type" : "object"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "pipeline-controller" ]
      }
    },
...
    "/api/portal/developer/provenance/requests" : {
      "get" : {
        "operationId" : "requests_1",
        "parameters" : [ {
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "default" : 30,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "type" : "string"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "provenance-controller" ]
      }
    },
...
    "/api/portal/developer/requests" : {
      "get" : {
        "operationId" : "requests",
        "parameters" : [ {
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "default" : 25,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/GatewayRequestLogEntity"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "portal-developer-controller" ]
      }
    },
...
    "/api/portal/developer/v8/firewall/{action}" : {
      "post" : {
        "operationId" : "firewall",
        "parameters" : [ {
          "in" : "path",
          "name" : "action",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : {
                    "type" : "object"
                  },
                  "type" : "object"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "v-8-controller" ]
      }
    },
...
  }
}